backend/requirements.txt

Django==3.1.2
django-cors-headers==3.5.0

django-widget-tweaks==1.4.8
djangorestframework==3.11.0
gevent==20.5.0
greenlet==0.4.15
gunicorn==20.0.4

psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1

